Put ListElement objects directly under parent lists
It was the case that nested list objects were placed under the preceding
ListItemElement in the structure:

  Document
    ListElement
      ListItemElement
      ListItemElement
        ListElement
         ListItemElement
      ListItemElement

This is probably how the resulting HTML will be structured, but this is
potentially specific to HTML output, and not really indicative of the
document's structure.  The HTML emitter will have to deal with this
special case in the structure, but that's a small price to pay.