Allow honing in on a single test
To hone in on a test on a particular line (must be the line where
test_doc is invoked), specify the following to prove:

    prove -l $test_file :: +89 # 89 is the example line number

To hone in on a test with a particular name (tests without names
are *always* ignored here), specify the following to prove:

    prove -l $test_file :: 'Test a basic list of unordered items'

The extra argument is treated as a regular expression.

Multiple specifications may be provided to run multiple tests.